wip: refactors more logic into player
Mostly dealing with a bug when killing bullets. it seems if the timer
has a reference to an inst that is freed, the whole game crashes badly.

Capturing that reference via scoping or an atom leads to a crash if it
becomes invalid....? Curious if deleting from clj instead of gd script
makes a difference.